+++ src/te/operation/scan_op.cc
@@ -21,6 +21,7 @@
  * \brief Scan Operator.
  * \file scan_op.cc
  */
+#include <tvm/runtime/registry.h>
 #include <tvm/te/operation.h>
 #include <tvm/tir/expr.h>
 #include <tvm/tir/ir_pass.h>
@@ -119,6 +120,10 @@ Operation ScanOpNode::make(std::string name,
   n->inputs = std::move(inputs);
   return Operation(n);
 }
+
+TVM_REGISTER_GLOBAL("te.ScanOp")
+.set_body_typed(ScanOpNode::make);
+
 
 Array<Tensor> scan(Array<Tensor> init,
                    Array<Tensor> update,
